Restart with chunk decoding functions
I made some previous assumptions about how the gen_rtmp
module (originally built in a different package) would work. The
assumption would be that we would receive as much data as possible from
the socket and then have the gen_rtmp process keep track of the unused
bytes and continue to process until it needs to get more. This time
around when implementing gen_rtmp, it is being "smart" enough to grab
the number of bytes it actually needs from the socket just as they are
needed.